"Fixed" shape mismatch in model with scalar observ and indexing by s
Fixed is, possibly, a strong word. I figured it was due to the plate.
May be a good idea to ask why it happens. Key concern is that under the
enumeration interpretation ``beta`` for some reason happen to acquire
3-4 singleton dims (depending on which dim was ascribed to the
``vectorized_markov`` plate) on the right. I don't see why it need to
happen. Indicating ``beta`` as deponent helps, but I am not sure if it
will be sufficient in the case of vector-values observations